requirements.txt
streamlit
numpy
scikit-image
scikit-learn
joblib
Pillow
